backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.14.0

# Settings & validation
pydantic-settings>=2.6.0

# HTTP client
httpx>=0.28.0

# Data & analysis
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0

# ML
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.4.0
torch>=2.4.0

# NLP & sentiment
transformers>=4.40.0
praw>=7.7.0

# LLM
langchain>=0.3.0
langchain-community>=0.3.0

# Scheduling & notifications
apscheduler>=3.10.0
resend>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Linting
ruff>=0.8.0
